Quitar el limite maximo de rondas y actualizar documentacion
- Solo se exige mas de 200 rondas (se elimina MAX_RONDAS)
- Historiales con new/delete[] del tamano exacto
- Puntajes y contadores en long long para evitar desbordes
- Vengativa mira solo la ronda anterior (mismo resultado, mas rapida)
- Reporte y guia del codigo (PDF y fuente HTML), diagramas y pruebas actualizados

+++ partida_iterada.h
@@ -11,6 +11,6 @@
 // Devuelve en puntosPartidaA / puntosPartidaB los puntos de cada una en esta
 // partida, y suma a cada estrategia las veces que eligio C y T.
 void ejecutarPartida(Estrategia &estrategiaA, Estrategia &estrategiaB, int numeroRondas,
-                     int &puntosPartidaA, int &puntosPartidaB);
+                     long long &puntosPartidaA, long long &puntosPartidaB);
 
 #endif // PARTIDA_ITERADA_H
